[cmake] check-security => security-check
Summary:
All the check-whatever targets are unit tests with the exception of check-security and check-symbol (which I intend to port over too). The script is named security-check.py and the function called by cmake as well. This is just more consistent all over this way.

Bonus: Use the COMMENT facility rather than echoing stuff on the command line when runnign the checks.

Test Plan:
  ninja all check security-check

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5441