Rebuild native executable when changes to the build system are made
Summary:
When the build rules needs to be updated, cmake isn't always rerun for the native build.

This adds a dependency file that ensures it does. This feature is limited to cmake, so there are no changes with other build systems and a call to clean might still be necessary.

Test Plan:
  ninja
  ninja # Check native build isn't ran again.
  touch build.ninja
  ninja # Native build runs again.

Build with make as a generator and verify it's still good.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D5624