Remove unnecessary sign variable from wnaf_const
Summary:
 * Fix test_constant_wnaf for -1 and add a test for it.

Before, test_constant_wnaf used scalar_cadd_bit to correct for the skew. But
this function does not correctly deal with overflows which is why num = -1
couldn't be tested.

This commit also adds tests for 0, 1/2 and 1/2-1 as they are corner cases
in constant_wnaf.

 * Remove unnecessary sign variable from wnaf_const

This is a backport of libsecp256k1 [[https://github.com/bitcoin-core/secp256k1/pull/741 | PR741]]

Test Plan:
  ninja check-secp256k1

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7595