[CMAKE] Better debug install targets
Summary:
The current `install-debug` behavior installs the same content than the
`install-all` target (then splits debug from it), which is not
intuitive. This diff renames it to `install-all-debug` and adds an
`install-debug` target that matches the `install` default target
instead.

Depends on D7240.

Test Plan:
  ninja install-debug
  ninja install-all-debug
Check the output are as expected

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7243