[LINTER] Add a YAML linter
Summary:
This uses a relaxed rule set for now, which does not enforce a strict
number of indentation spaces but instead check the file is consistently
indented. The maximum line lenght rule is also disabled to accomodate
the scripts in the gitian descriptors.

Test Plan:
Should return no error:
  arc lint --everything
Revert the changes to the .travis.yml from this diff then run:
  arc lint
and check the indentation errors are catched by the linter.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7358