[CMAKE] Generate the man pages at install time
Summary:
The man pages are currently part of our source tree and maintained as
such despite being generated. This diff removes the man pages from the
source tree and generate them as part of the intallation step. The man
pages are still part of the release delivery.

Test Plan:
  ninja install-manpages
Check the man pages are installed.
  touch ../src/bitcoind.cpp
  ninja install-manpages
Check the man pages are regenerated.
  touch ../src/bitcoin-cli.cpp
  ninja install-manpages
Check only the bitcoin-cli man page is regenerated.
  ninja install-manpages-html
Check the html variants of the manpages are installed.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7446