[CMAKE] Exclude the man pages from the install-debug target
Summary:
The man pages are now generated, and no longer part of the default
install target. This diff forwards the exclusion to the install-debug
target, since there is nothing to strip in a man page.

This has the side effect of removing the manpages from the Linux and
Windows Gitian builds, which is a good thing because the generation
process involves running the binary to extract the help, and this is not
always possible when cross compiling. The diff also removes the manpages
from the OSX Gitian build, for the same reason.

Test Plan:
  ninja install-debug
Check the manpages are not installed.
Run the gitian builds.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7452