[CMAKE] Move the manpages generation logic to doc/man
Summary:
The man pages generation is not part of the installation process, so
move it to `doc/man` where it belongs.
The `install_manpage()` interface now takes a list of pages to install, which can be empty.
This allows for adding the pages later by filling the `MAN_PAGES` target property.

Test Plan:
  ninja install-manpages

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7473